storage: extract metrics-related code
Release note: None
